--- pages/components/Image.tsx
@@ -1,49 +1,0 @@
-import React from 'react';
-
-const Image = () => {
-    return (
-        <div className='box-border overflow-hidden pt-24 w-full gap-0 columns-3'>
-            <div className=" bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png')]">
-                The Perfume
-            </div>
-            <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('https://miro.medium.com/max/400/1*X5-pQC_yDOuSMPlrXGporw.png')]">
-                Bumblebee Icons
-            </div>
-            <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png')]">
-            Häng konsten lågt
-            </div>
-            <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('https://miro.medium.com/max/400/1*X5-pQC_yDOuSMPlrXGporw.png')]">
-            We are the shit 
-            </div>
-            <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png')]">
-            Candy Icons 
-            </div>
-            <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('https://miro.medium.com/max/400/1*X5-pQC_yDOuSMPlrXGporw.png')]">
-            Ford Keyless
-            </div>
-            <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png')]">
-            Think
-            </div>
-            <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('https://miro.medium.com/max/400/1*X5-pQC_yDOuSMPlrXGporw.png')]">
-            Malmö Festival
-            </div>
-            <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png')]">
-            La Bitácora de Viajero
-            </div>
-        </div>
-    );
-};
-
-export default Image;
-
-{/* <div className='box-border overflow-hidden pt-24 w-full gap-0 columns-3'>
-<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full" src="https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png" alt="" /></div>
-<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://miro.medium.com/max/400/1*X5-pQC_yDOuSMPlrXGporw.png" alt="" /></div>
-<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png" alt="" /></div>
-<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://miro.medium.com/max/400/1*X5-pQC_yDOuSMPlrXGporw.png" alt="" /></div>
-<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png" alt="" /></div>
-<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://miro.medium.com/max/400/1*X5-pQC_yDOuSMPlrXGporw.png" alt="" /></div>
-<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png" alt="" /></div>
-<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://miro.medium.com/max/400/1*X5-pQC_yDOuSMPlrXGporw.png" alt="" /></div>
-<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png" alt="" /></div>
-</div> */}
