+++ pages/components/Images.tsx
@@ -1,0 +1,69 @@
+import React from "react";
+
+const Image = () => {
+  return (
+    <div className="box-border overflow-hidden pt-24 w-full gap-0 columns-3">
+      <div className=" bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat w-full bg-[url('http://core.loopthemes.com/wp-content/uploads/2016/04/Project1_2b-1280x720.jpg')]">
+        <div className="text-[3rem] font-serif text-yellow-600 opacity-0 hover:opacity-100 w-full text-center h-full justify-center flex items-center duration-500 bg-white/80 overflow-hidden">
+          The Perfume
+        </div>
+      </div>
+      <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('http://core.loopthemes.com/wp-content/uploads/2016/03/bumblebee.jpg')]">
+        <div className="text-center  text-[3rem] font-serif text-yellow-600 opacity-0 hover:opacity-100 w-full h-full justify-center flex items-center duration-500 bg-white/80 overflow-hidden">
+          Bumblebee Icons
+        </div>
+      </div>
+      <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('http://core.loopthemes.com/wp-content/uploads/2016/03/tuli_co-1024x1024.jpg')]">
+        <div className="text-center  text-[3rem] font-serif text-yellow-600 opacity-0 hover:opacity-100 w-full h-full justify-center flex items-center duration-500 bg-white/80 overflow-hidden">
+          Häng konsten lågt
+        </div>
+      </div>
+      <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('http://core.loopthemes.com/wp-content/uploads/2016/04/wearetheshit.jpg')]">
+        <div className="text-center  text-[3rem] font-serif text-yellow-600 opacity-0 hover:opacity-100 w-full h-full justify-center flex items-center duration-500 bg-white/80 overflow-hidden">
+          We are the shit
+        </div>
+      </div>
+      <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('http://core.loopthemes.com/wp-content/uploads/2016/04/af3c5019650315.562efb07ec97e.gif')]">
+        <div className="text-center  text-[3rem] font-serif text-yellow-600 opacity-0 hover:opacity-100 w-full h-full justify-center flex items-center duration-500 bg-white/80 overflow-hidden">
+          Candy Icons
+        </div>
+      </div>
+      <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('http://core.loopthemes.com/wp-content/uploads/2016/04/ford-keyless.jpg')]">
+        <div className="text-center  text-[3rem] font-serif text-yellow-600 opacity-0 hover:opacity-100 w-full h-full justify-center flex items-center duration-500 bg-white/80 overflow-hidden">
+          Ford Keyless
+        </div>
+      </div>
+      <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('http://core.loopthemes.com/wp-content/uploads/2016/04/think03-600x400.jpg')]">
+        <div className="text-center  text-[3rem] font-serif text-yellow-600 opacity-0 hover:opacity-100 w-full h-full justify-center flex items-center duration-500 bg-white/80 overflow-hidden">
+          Think
+        </div>
+      </div>
+      <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('http://core.loopthemes.com/wp-content/uploads/2016/04/malmo02-514x600.jpg')]">
+        <div className="text-center  text-[3rem] font-serif text-yellow-600 opacity-0 hover:opacity-100 w-full h-full justify-center flex items-center duration-500 bg-white/80 overflow-hidden">
+          Malmö Festival
+        </div>
+      </div>
+      <div className="bg-cover bg-center justify-center flex items-center aspect-square bg-no-repeat  w-full bg-[url('http://core.loopthemes.com/wp-content/uploads/2016/03/70afac33369027.56a9085d0079b-600x400.jpg')]">
+        <div className="text-center  text-[3rem] font-serif text-yellow-600 opacity-0 hover:opacity-100 w-full h-full justify-center flex items-center duration-500 bg-white/80 overflow-hidden">
+          La Bitácora de Viajero
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Image;
+
+{
+  /* <div className='box-border overflow-hidden pt-24 w-full gap-0 columns-3'>
+<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full" src="https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png" alt="" /></div>
+<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://miro.medium.com/max/400/1*X5-pQC_yDOuSMPlrXGporw.png" alt="" /></div>
+<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png" alt="" /></div>
+<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://miro.medium.com/max/400/1*X5-pQC_yDOuSMPlrXGporw.png" alt="" /></div>
+<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png" alt="" /></div>
+<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://miro.medium.com/max/400/1*X5-pQC_yDOuSMPlrXGporw.png" alt="" /></div>
+<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png" alt="" /></div>
+<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://miro.medium.com/max/400/1*X5-pQC_yDOuSMPlrXGporw.png" alt="" /></div>
+<div className=''><img className="duration-700 hover:sepia min-w-[50px] min-h-[50px] w-full " src="https://pbs.twimg.com/profile_images/446356636710363136/OYIaJ1KK_400x400.png" alt="" /></div>
+</div> */
+}
