+++ pages/components/Menu.tsx
@@ -3,16 +3,17 @@ import React from "react";
 const Menu = () => {
   return (
     <div>
-      <div className="">
-        <button className="text-black font-semibold rounded inline-flex float-right pt-14 pr-20 text-2xl">
+      <div id='nav.bar' className="sticky top-0">
+        <button className="text-black font-semibold inline-flex float-right pt-14 pr-20 text-2xl">
           MENU
         </button>
-        <button className="text-black font-semibold rounded inline-flex float-left pt-14 pl-20 text-2xl">
+        <button className="text-black font-semibold inline-flex float-left pt-14 pl-20 text-2xl">
           C O R E
         </button>
       </div>
     </div>
   );
 };
+
 
 export default Menu;
